m_dataSize has type std::size_t because unsigned int is too small

+++ include/table_types/ChebyInterpTable.hpp
@@ -55,7 +55,7 @@ public:
     m_name = grid_type_to_string<GT>() + "ChebyInterpTable<" + std::to_string(N) + ">";
     m_numTableEntries = m_numIntervals+1;
     m_order = N+1; // N is the degree of the polynomial interpolant so the order is N+1
-    m_dataSize = static_cast<unsigned>(sizeof(m_table[0]) * m_numTableEntries);
+    m_dataSize = sizeof(m_table[0]) * m_numTableEntries;
 
     auto fun = func_container.standard_fun;
